requirements.txt
tweepy
discord.py
requests
python-dotenv
aiohttp